crypto: sa2ul: fix dma_sync_sg_for_device usage
dma_sync_sg_for_device should use DMA device as the device, not the
sa2ul device. Syncing on the sa2ul device is unnecessary and rather
expensive computationally also for no purpose; sa2ul itself is never
going to access the data, it is only DMA who does it.

Reviewed-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Tero Kristo <t-kristo@ti.com>