crypto: sa2ul: avoid usage of sg_split under AEAD if possible
Using sg_split is a relatively heavy operation, so avoid using it if
possible. We can use a scatterlist entry stored locally within the
crypto context if our scatterlist only contains a single entry. In this
case, we just map the data pointers over to our local entry instead of
creating a complete split.

Reviewed-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Tero Kristo <t-kristo@ti.com>