iommu/omap: Add registration for DT fwnode pointer
[ Upstream commit 5df362a53f7d36e032668e7e6725d80622b98525 ]

The fwnode pointer must be passed to the iommu core, so that the core
can map the IOMMU towards device requests properly. Without this, some
IOMMU clients like OMAP remoteproc will fail the iommu configuration
multiple times with -EPROBE_DEFER, which will eventually be ignored with
a kernel warning banner.

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Link: https://lore.kernel.org/r/20200424145828.3159-1-t-kristo@ti.com
Signed-off-by: Joerg Roedel <jroedel@suse.de>
[s-anna@ti.com: cherry-pick commit '5df362a53f7d' from v5.8]
Signed-off-by: Suman Anna <s-anna@ti.com>