remoteproc/pru: Cleanup of_device_is_compatible() usage
The PRU remoteproc driver uses the of_device_is_compatible() function
during probe to dynamically assign some flags and properties for each
PRU core. This usage is not recommended and makes the code a bit
cumbersome. Cleanup most of this usage by using device compatible
match data. The check for K2G to conditionally avoid the mailbox
usage is the only check left-out.

Signed-off-by: Suman Anna <s-anna@ti.com>