arm64: dts: ti: am654-base-board: Add a board dts variant for SR1.0
The AM65x family of SoCs has two Silicon Revisions - SR1.0 and SR2.0.
The AM65x SR2.0 will be the officially supported revision, and all
the current dtsi and dts files will be reused to represent this Silicon
Revision.

The AM65x SR2.0 SoCs have a revised ICSSG IP that is based off the
subsequent IP revision used on J721E SoCs. The ICSSG IP on AM65x SR2.0
SoCs have two new custom auxiliary PRU cores called Transmit PRUs
(Tx_PRUs) in addition to the existing PRUs and RTUs, but these are
not present on AM65x SR1.0 SoCs. The Tx_PRU nodes are added and enabled
by default in the base k3-am654-main.dtsi file.

Add a new board dts variant file am654-base-board-sr1.dts, that inherits
the regular am654-base-board.dts file, and will be customized to add
any changes required to boot a AM65x board using the SR1.0 Silicon.
The Tx_PRU nodes are absent on SR1.0 and so are disabled to begin
with.

NOTE:
This will not be upstreamed, and so there is no refactoring done to
either introduce a am654-base-board-common.dtsi or a am65x-sr1.dtsi
files.

Signed-off-by: Suman Anna <s-anna@ti.com>