arm64: dts: ti: k3-j721e-main: Configure MAIN R5FSS1 for Split-mode
Switch the MAIN R5FSS1 cluster to be configured for Split-mode as the
default so that two different applications can be run on each of the
R5F cores in performance mode. LockStep-mode would be available only
on SoCs efused with the appropriate capability bit, and Split-mode is
the mode that is available on all J721E SoCs.

Signed-off-by: Suman Anna <s-anna@ti.com>