net: ethernet: ti: icssg_prueth: Add shadow mode support to IEP
This is to prepare for SR2.0 support which uses the
IEP in 32-bit shadow mode instead of 64-bit mode.

If cycle_time_ns parameter is non zero then the IEP
will be setup in 32-bit shadow mode.

As existing set/get/adjust time operations are not
suitable for SR2.0 add 'struct icssg_iep_clockops' to
provide alternate implementation for those functions.

Signed-off-by: Roger Quadros <rogerq@ti.com>