net: ethernet: ti: icssg_prueth: Support SR2.0
A number of things have changed in SR2.0 firmware.
This patch addresses those while maintaining compatibility
with SR1.0 firmware.

- TX_PRU core added
- Config mechanism has changed. ICSSG Hardware queue module
- DMEM registers are used to provide initial configuration.
- ICSSG Hardwre queue module is used to send commands to firmware
and obtain the TX timestamp. Configuration over management channel
is not required.
- IEP made optional as now only 1 port can use IEP.
- IEP operation has changed. 32-bit shadow mode is used instead
of 64-bit mode. DMEM registers needs to be used to calculate
ns timestamp along with IEP count register.

For SR1.0 the only change is that IEP node is now optional
for each port.

Signed-off-by: Roger Quadros <rogerq@ti.com>