net: hsr/prp: add procfs interface for snmp agent interface
implement a simple procfs interface to support snmpd agent
set/get command to implement iec62439 mibs for LRE. This is done
similar to ip/udp/tcp mibs of mibii for stats and such. procfs
files are created under the /proc/<interface-name> folder.

This patch also add some missing hsr/prp modes so that same can
be set through SNMP interface.

The format of print to console is kept same as the current debugfs
output format expected by snmpd so that same text parser can be
re-used at snmpd even after this migration.

Note that this is a temporary interface and is not accepted by
netdev community at LKML and may change in the future.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Reviewed-by: Grygorii Strashko <grygorii.strashko@ti.com>