Enable EFI so that standard UEFI boot loaders can be used.
On ARM it is usually U-Boot that acts as a UEFI provider, and both
U-Boot and GRUB can act as UEFI apps. However, for them to load the
kernel, the kernel must have EFI enabled.

Although this may not be used by Beagleboard itself, it is used by
downstream projects, such as Mender, to provide standardized boot and
rollback methods.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>