arm: dts: am335x: add pruss uart controller node
The PRUSS on AM335x has a UART sub-module which is based on the industry
standard TL16C550 UART controller, supporting 16x and 13x over
samplings. The UART input clock frequency is 192MHz.

The node is added to the common am33xx-l4.dtsi file and is disabled. The
node needs to be enabled in the respective board files which uses
relevant AM335x SoCs supporting PRUSS, along with required PRU
application node properties.

Signed-off-by: Bin Liu <b-liu@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>