Revert "mtd: spi-nor: cadence-quadspi: Enter/exit DDR mode as required"
This reverts commit d3d4f00163b825c47e9ff01ccd299ba03f76a3d9.

The way DDR mode is activated is being changed in SPI NOR. It will be
done with SPI MEM as the main focus. To do that, the old way of
activating DDR mode is being replaced with one that is proposed
upstream. Drop this change to prevent the build from failing because
things like SPI_NOR_MODE_OPI_DTR are removed from the SPI NOR core.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>