net: hsr/prp: add vlan ctag filter support
This patch adds support for vlan ctag based filtering at slave devices.
The slave Ethenet device may be capable of filtering Ethernet packets
based on VLAN ID. This requires that when the VLAN interface is created over
an hsr/prp interface, it passes the vid information to the associated
slave Ethernet devices so that it updates the hardware filters to
filter Ethernet frames based on VID. This patch adds the required
functions to propagate the vid information to the slave devices.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Reviewed-by: Grygorii Strashko <grygorii.strashko@ti.com>