net: ti: prueth_core: move timer utility functions to prueth_core
As a preparatory patch to introduce additional Ethernet types in
the driver that also requires timer related functions, it is cleaner
to have timer related main code in prueth_core driver and on event
make calls to appropriate firmware specific functions as needed. So
this add a function prueth_enable_nsp(), to configure storm prevention
and define the same in prueth_qos.c. On timeout, call prueth_enable_nsp()
to configure storm prevention feature.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>