net: ti: prueth: dual_emac: Add support for timestamping tx packets
When PTP support is enabled in dual emac firmware, the firmware timestamps
the ptp packets that are going out via prueth ports. The timestamp value
is copied to a shared memory location. After timestamping, a host irq is
raised by the firmware as a notification to read the timestamp.

Add support for handling this ptp tx irq, read the tx timestamp and pass
it to userspace layer.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>