net: ti: prueth_core: add helper functions to support lre offload
Firmware requires node table entries to be inserted by driver to save
PRU cycles for packet processing. Node table provides a view of remote
nodes in the network. The nodes send supervision frames periodically
with its own mac address and the same is processed by firmware. First
firmware lookup the source MAC address. If not found, it indicates a
flag in the descriptor which driver uses to insert the MAC address
in the node table. The rx packet processing code kicks a worker thread
to insert the mac address. This patch adds the required function to
do the insert function and also clear the table on demand.

Firmware also requires driver to run a timer for 10 msec to update
check flags in dram for firmware to do periodic table maintenance
functions for LRE offload. So change basic timer period to 10 msec
and use a counter to track the NSP timer which needs to be done every
100 msec.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>