net: ti: prueth_core: Fix initialization of iep timer
IEP is initialized during the probe of prueth driver. When switching to

different firmware(from dual emac to HSR) IEP counter is not reset and
current timer is used. Also when both the ports are down, IEP need not
be running as there are no users. So, move icss_iep_init/exit() calls to
emac_ndo_open/stop() respectively.

Fixes: 5ed1a006f42a ("net: ti: prueth_core: Add support for initializing IEP")
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>