ARM: dts: am437x: Add PRUSS eCAP node
The PRUSS on AM437x SoCs has an Enhanced Capture (eCAP) sub-module
that can be used by PRU Ethernet to provide rx interrupt pacing
support.

Add the dt node for this eCAP module as a child node within the
PRU-ICSS1 node. The eCAP module is not pinned out on PRU-ICSS0
instnace.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>