net: ti: prueth: Use the ethertype in skb_redundant_info to detect hsr/prp
prueth_ptp_ts_event_type is using eth_hdr(skb)->h_proto type to detect
hsr and prp. But with the VLAN tag the ethernet type is not populated
correctly. The ethertype in skb_redundant_info is populated correctly by
the hsr core layer. Use it for detecting hsr and prp type.

Fixes: 3b9b5f341341 ("net: ti: prueth: hsr: Add support for timestamping tx packets")
Reviewed-by: Murali Karicheri <m-karicheri2@ti.com>
Reviewed-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>