remoteproc/k3-dsp: Remove unneeded pm_runtime usage
The TI K3 DSP remoteproc driver uses couple of pm_runtime API in probe
and remove to turn on and off the necessary clocks associated with the
DSP internal memories. However, these are no-ops for this driver, as
the DTS nodes do not use any power-domains, and so there is no real
back-end behind these calls. These clocks are instead turned on and
off during the .prepare() and .unprepare() callbacks. So, eliminate
the dead pm_runtime API usage from the driver.

Signed-off-by: Suman Anna <s-anna@ti.com>