remoteproc/k3-dsp: Use device-specific match data
The K3 DSP remoteproc driver is used to boot two different classes
of DSPs - a 32-bit C66x DSP and a 64-bit C71x DSP. There are couple
of differences in the features managed by the driver between the
two DSP classes - number of internal RAMs and their addresses, the
local reset integration, and the boot vector address alignment. These
are currently handled in code.

Introduce and use device-specific data which is retrieved using the
compatible match data logic. Following are the main summary of changes:
 - The k3_dsp_rproc_of_get_memories() is simplified to remove the
   hard-coded string names and compatible-check logic to skip over
   irrelevant memories; and the DSP address view for each of the
   internal memories.
 - Add the boot vector alignment to the device data and use it to
   perform a sanity-check of the boot address in k3_dsp_rproc_start().
   The added logic introduces the correct alignment check for C71x
   cores.
 - The uses_lreset is also moved from the remoteproc strucutre to the
   device-specific data, which is cached in the k3_dsp_rproc structure.
 - The structure k3_dsp_rproc_mem is renamed to k3_dsp_mem.

Signed-off-by: Suman Anna <s-anna@ti.com>