spi: spi-mem: allow telling controllers the op to use for calibration
Some controllers like the Cadence QSPI controller need to perform a
calibration sequence to operate at higher frequencies. This calibration
involves reading a known tuning pattern from the flash.

To be able to read this, the controller needs to know the opcode for
reads, the dummy cycles needed, etc. So, allow SPI NOR to tell the
controller the read op template that can be used for calibration.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>