TEMP: arm64: dts: ti: k3-j7200: Add aliases for R5F rproc nodes
Add aliases for all the R5F remoteproc processor cores present
within the MCU domain and MAIN domain and common to all the K3
J7200 boards. The aliases uses the stem "rproc". Note that each
of the R5F1 cores and their corresponding aliases are uniquely
identified and used only if the corresponding R5F cluster/subsystem
is running in Split (non-LockStep) mode.

The aliases can be overridden, if needed, in the respective
derivative board dts files.

Signed-off-by: Suman Anna <s-anna@ti.com>