Revert "arm64: dts: ti: j721e-main: add timesync_router node"
This reverts commit dea1b0d4e67a8de1e8a51e31433c92192b056a06.

The TimeSync Interrupt Router module is currently represented as
a pinctrl device in Linux, with the individual pinmux settings
configuring the interrupt router connections. This IP is an
Interrupt Router and is in general firewalled away and needs
to be programmed using System Firmware.

Remove this pinctrl device in preparation to rework and model
it as a proper TI-SCI Interrupt Router device.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>