net: ethernet: ti: prueth: Fix a race condition in passing cut through packets
When firmware sends the cut through packet to other port, it is possible
that host receives the packet first. In this case, host receives a tx
interrupt with cut through ts stored. However driver is pushing the
packet to user space without storing the timestamp. Fix it by storing the
ts before pushing the skb to user space.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>