dt-bindings: hwlock: omap: Convert binding to YAML
[ Upstream commit d8db9dc348712bbe972e1edd952b67279e774ecc ]

Convert the current OMAP hwspinlock binding from text format to YAML
format/DT schema, and delete the legacy text binding file.

The new YAML binding conversion is a slightly updated version compared
to the original. The legacy "ti,hwmods" property is now obsolete and
is dropped altogether, and the K3 example is updated to showcase the
actual dts node usage.

[s-anna@ti.com: cherry-pick commit 'd8db9dc34871' staged for v5.10]
Signed-off-by: Suman Anna <s-anna@ti.com>
Link: https://lore.kernel.org/r/20200828041447.5900-1-s-anna@ti.com
Signed-off-by: Rob Herring <robh@kernel.org>