net: ethernet: ti: icssg_prueth: add multiq support for SR2
In the Tx/Egress direction, ICSSG firmware has 8 internal strict
priority queues that are used to queue the frames received from the
Host. However there are only 4 PSI threads/channels from host to ICSSG
which are equivalent to hw queue definition of Linux. So implement 4
hw queues in the Egress direction using netdev multiq framework for
SR2 of ICSSG.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Acked-by: Roger Quadros <rogerq@ti.com>