net: ethernet: ti: icssg_prueth: cleanup TX/RX channels on error
Currently SKBs queued to RX channels are not freed on error in the
subsequent code in emac_ndo_open(). Similarly TX channels are not
disabled on error in subsequent code. This patch introduces two
cleanup functions, prueth_reset_tx_chan() and prueth_reset_rx_chan()
and use it from emac_ndo_stop() as well as from the error path to
clean up the channels.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Acked-by: Roger Quadros <rogerq@ti.com>