rpmsg: char: Add device id_table for auto-probe
The rpmsg char driver is not auto-probed for any rpmsg device
without relying on the device's driver_override feature. Add a
device id_table to the rpmsg-char driver to facilitate this
auto-probe. Any rpmsg device with the name "rpmsg_chrdev" will
be auto-probed to begin with.

The driver_override will continue to work just fine as well.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>