net: phy: Add a helper to return the index for of the internal delay
commit 92252eec913b2dd5e7b5de11ea3efa2e64d65cf4 upstream.

Add a helper function that will return the index in the array for the
passed in internal delay value.  The helper requires the array, size and
delay value.

The helper will then return the index for the exact match or return the
index for the index to the closest smaller value.

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>