net: ethernet: ti: icssg_prueth: Add MAC address to FT1 slot 0
This is required by firmware to deal with Source Address Violation
for zero address for example if one of the ports has MAC address
set to 0 (e.g. when only one port was brought up).

Without this f/w drops packets containing zero address even if
promiscuous mode is enabled.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>