net: ethernet: ti: prueth: Remove IRQF_ONESHOT when requesting IRQ
IRQF_ONESHOT is explicitly defined to be used with request_threaded_irq()
when threaded handler is provided. Also it blocks IRQ handler to become
threaded on Linux RT.

One choice was to use IRQF_NO_THREAD to reatain the hardirq
context if required but this will most likely be NACked for upstream
as was done for CPSW driver.

Remove IRQF_ONESHOT usage and drop 'hard' from handler names as they no
longer will be run in hardirq context.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>