arm64: dts: ti: k3-am64: Add aliases for R5F rproc nodes
Add aliases for all the R5F remoteproc processor cores present
within the MAIN domain and common to all the K3 AM64x boards.
The aliases use the stem "rproc". Note that each of the R5F1
cores and their corresponding aliases are uniquely identified
and used only if the corresponding R5F cluster/subsystem is
running in Split-mode.

The aliases can be overridden, if needed, in the respective
derivative board dts files.

Signed-off-by: Suman Anna <s-anna@ti.com>