spi: cadence-quadspi: Check if the pattern is present before tuning PHY
If the tuning pattern is not present then the tuning will fail at the
first step anyway, but it will emit a warning message like "PHY
calibration failed: -2". This might give users the wrong impression that
something fatal is wrong with the flash. It is only a soft error. The
read speeds will be slower but everything should still work. So check if
the pattern is present on the flash before running the tuning algorithm.

While here, mention that PHY calibration failure would result in slower
read speeds in the case where the tuning pattern _is_ present and bump
it up to a warning. This will clarify to users what exactly will happen
as a result of the calibration failure.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>