dt-bindings: usb: Fix USB dt-bindings for AM64 SoC
AnyOf will only let any one of the compatible strings to be used in a DT
node. Also, dt_binding_check fails because '\t' is used instead of spaces
at the start of line.

Fix it by using oneOf property to list the possible ways compatible strings
can be used in a device tree node.

Fixes: 1323070118b5 ("dt-bindings: usb: Add new compatible string for AM64 SoC")
Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>