net: ethernet: ti: icss_iep: Fix reading of capture registers
caputre events in icss_iep_cap_cmp_handler() is compared against variable
val containing the capture status register value. But this variable is
re used later in the function. Because of this unnecessary external ts
event are received in userspace. Fix it by using a sts variable for
storing capture status register.

Fixes: c1df31d84264 ("net: ethernet: ti: icss-iep: Add support for generating extts event")
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>