arm64: dts: ti: k3-am642-sk: Reserve memory for IPC between RTOS cores
Add a reserved memory node to reserve a portion of the DDR memory to be
used for performing inter-processor communication between all the remote
processors running RTOS or baremetal on the TI K3 AM642 SK board. 8 MB of
memory is reserved for this purpose, and this accounts for all the vrings
and vring buffers between all the possible pairs of remote processors.

The reserved memory set aside is the same as the respective carveout on
the AM64x EVM board to maintain firmware compatibility between the two
boards. Please see commit bfe196803848 ("arm64: dts: ti: k3-am642-evm:
Reserve memory for IPC between RTOS cores").

Tested-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>