ti_config_fragments: rpmsg/v8_rpmsg: Enable rpmsg_char driver
Enable the generic rpmsg_char driver (RPMSG_CHAR) in both the
ipc.cfg and v8_ipc.cfg so that the driver is built as a module
by default on all applicable TI platforms. The module is built
as a rpmsg_char.ko, and the driver name in sysfs is rpmsg_chrdev.

The driver presents a /dev/rpmsg_ctrlX (where X is an allocated
number) for each rpmsg device that can probe this driver. The
driver currently does not get auto-probed, but if desired, can
have the published rpmsg devices be bound to this driver using
udev rules or some userspace admin process.

Signed-off-by: Suman Anna <s-anna@ti.com>