rpmsg-kdrv: Add support for device virtualization
In J721E, multiple compute entities can run in parallel,
and can require to use any given H/W simultaneously.

Or, a device may be scheduled between multiple cores, and
a dedicated core can manage the switching based on
requests / policies

rpmsg-kdrv is a device virtualization framework that uses
rpmsg core functionalities. In a system using this feature,
a remoteproc is expected to advertise virtual device
endpoints that the host can connect to and request for
services.

Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Tested-by: Grygorii Strashko <grygorii.strashko@ti.com>