rpmsg: rpc: add the uapi rpmsg_rpc header
A new rpmsg client driver, rpmsg_rpc, will be introduced later that will
provide a framework for userspace applications to execute functions on
different remote processors.

The driver added in previous downstream TI kernels is not yet ready and
requires rework due to changes in dma-buf framework, to be ported and
be functional properly onto 5.10 kernel. Add and export the required
minimal uapi rpmsg_rpc.h header file for now to unblock the integration
of the TI IPC userspace components.

Signed-off-by: Suman Anna <s-anna@ti.com>