net: ti: prueth_core: Add support for timestamping tx/rx packets using IEP
IEP inside ICSS is used as PTP clock. Add support for initializing IEP
clock when prueth is enabled on ICSS. When PTP support is enabled, the
firmware timestamps the PTP packets that are going out via prueth ports.
The timestamp value is copied to a shared memory location. After
timestamping, a host IRQ is raised by the firmware as a notification to
read the timestamp. When PTP packets are received on ICSS ports,
firmware record the timestamps and appends the timestamp at
the next 32 byte block after the end

Add support to read the tx/rx timestamp and pass it to userspace.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>