arm: dts: add am335x-boneblack-pruswuart.dts
This adds new dts file am335x-boneblack-pruswuart.dts to support PRU
software-based UART on Beaglebone Black.

The UART pins are on the expansion connector P8 and P9. Showing as an
example, the 3 ports on PRU0 do not define cts/rts pins to not use hw
flow control, while the 3 ports on PRU1 define cts/rts pins.

Signed-off-by: Bin Liu <b-liu@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>