arm64: dts: ti: am654: Add a SR1.0 board dts overlay and disable UHS modes
The 4 bit interface in AM65x SR1.0 cannot be supported at 3.3V or 1.8V
because of erratas i2025 and i2026 [1]. However, the SD card is the
primary boot mode for development usecases.

Continue to enable SD card in SR1.0 and disable UHS modes in it to
minimise any ageing issues happening because of the erratas.

Add a new board dts overlay file 'k3-am654-base-board-sr1.dts' that
inherits the generic SR1.0 overlay 'k3-am654-sr1.dts' file, and disable
the UHS modes on the MMC sdhci1 node. This new overlay file is designed
to apply directly on top of the regular k3-am654-base-board.dtb file,
and requires corresponding bootloader support.

[1] https://www.ti.com/lit/er/sprz452e/sprz452e.pdf

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>
Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>