dt-bindings: net: Add binding for ti,icssg-prueth device
This is the DT binding document for the ICSSG Dual-EMAC Ethernet
device.

SR2.0 support:
We need to specify PRU handle and firmware for TX_PRU, configure
ICSS gp-mux in Linux and provide TX timestamp IRQ.
IEP is now optional for the emac nodes.
Also update the example to reflect SR2.0 case.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>