arm64: dts: ti: k3-am65-main: fix DSS irq trigger type
DSS irq trigger type is set to IRQ_TYPE_EDGE_RISING. For some reason this
results in double the amount of expected interrupts, e.g. for normal
page flipping test the number of interrupts per second is 2 * fps. It is
as if the IRQ would trigger on both edges.

In any case, it's better to use IRQ_TYPE_LEVEL_HIGH, like the other
devices do, which seems to fix this problem.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Tested-by: Praneeth Bajjuri <praneeth@ti.com>