ti_config_fragments: v8_baseport/debug: Enable EFI
Enable EFI stub defaults so that standard distros shall start working
with TI kernel configurations. This allows "bootefi" to finally start
working from u-boot and EFI compatible bootloaders. This ofcourse, is in
addition to current scheme of booting via "booti" u-boot commands.

Tradeoff is additional size (a 5.10 view of bloat-o-meter reports):
add/remove: 544/0 grow/shrink: 42/2 up/down: 114519/-184 (114335)

If this turns out to be a concern, we can consider moving this
as a separate config fragment of it's own.

Reported-by: Tom Rini <trini@konsulko.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Lokesh Vutla <lokeshvutla@ti.com>