net: ethernet: icssg-prueth: enable "fixed-link" connection
The ICSSG has incomplete and incorrect "fixed-link" connection type
implementation. The "fixed-link" is property of the Ethernet port and not
provided in "phy-handle" property.

So fix it by first checking for real PHY (phy-handle) and then for child
"fixed-link" property. After this patch the below DT definition should
allow to use fixed-link connection:

          icssg2_emac1: ethernet-mii1 {
                 phy-mode = "rgmii-rxid";
                 fixed-link {
                        speed = <1000>;
                        full-duplex;
                 };
          };

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>