arm64: dts: ti: k3-am64-main: Move TF-A location to 0x701c0000 and split the DMSC region
Earlier, the region 0x701c0000 to 0x701dffff was firewalled off because of
a bug in SYSFW. In the v2021.05 release of SYSFW this bug has been fixed
and this region can now be used for other allocations.

Therefore, TF-A's location has been moved to 0x701c0000 in ti-u-boot 08.00.

Also, split the DMSC region into DMSC-lite code region and secure proxy
communication buffer region. The memory region used for DMSC-lite code can
be optionally freed up by secure firmware API[1]. However, the buffer for
secure proxy communication is not configurable. This default hardware
configuration is unique for AM64.

[1] - http://downloads.ti.com/tisci/esd/latest/6_topic_user_guides/security_handover.html#triggering-security-handover

Fixes: 116621059eb1 ("arm64: dts: ti: k3-am64-main: Add SYSFW reserved ranges in OCRAM")
Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>