net: hsr: Support VLAN tagged supervision frames
In hsr/prp networks, user may choose to use vlan tagged Supervision (SV)
frames in the network which is defined as an optional feature in the
iec-62439 standard. User would then require to configure the VLAN ID,
PCP and DEI values to be used in the SV frames. This patch update the
driver to accept these values optionally from user as part of the
ip link add command and use VLAN info to format the VLAN tag in the
SV frame. Also pass the VLAN ID to the lower device so that it may
be added to the VLAN filter in h/w. Update the procfs interface to
allow user to enable/disable SV frame generation logic for debug
purpose.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>