net: hsr: Add SV frame counts in the rx direction
Add rx Supervision (SV) frame count for Link A and B to
debug stats. This is needed as the LRE rx stats is inclusive
of SV frames and user may want to count how many rx non SV
frames received for debug purpose. So add these counters to
the lre_info debugfs and update it in the packet processing
code.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>