net: ti: prueth_core: hsr/prp: set mux_sel and Min/Max frame sizes
As an preparatory patch to support other Ethernet types in the driver,
driver needs to configure min/max frame sizes in MII_RT registers as
well as MUX_SEL as part of prueth_mii_init(). It is expected that this
function gets called in ndo_open() and Ethernet type may be different
than EMAC or SWITCH.  So update prueth_mii_init() to configure the
values for High Seamless Redundancy (HSR) and Parallel Redundancy
Protocol (PRP) Ethernet types. HSR and PRP are two industrial
protocols that operates at Layer 2 to implement redundancy. They
extend the frame size by 6 octets of protocol bytes at Layer 2.
Thus update the max frame size for these Ethernet types in
MII_RT_RX_FRMS. As standard management frames are expected to be
received as well, min frame size remains unaffected.
Introduce a new macro PRUETH_IS_LRE() to check for HSR or PRP
and use that to configure the Max frame value differently. Link
Redundancy Entity (LRE) is the common function in HSR and PRP that
implements redundancy for HSR and PRP. Also MUX_SEL reg values
are same for HSR, PRP and SWITCH. So adjust the code accordingly.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>