Revert "spi: cadence-quadspi: do not use DMA for reads smaller than 16 bytes"
This reverts commit c9293d064e1103f630018efca18caf4f1ea93658.

This commit causes a glitch with register writes on J7200. As a result
the dummy cycle configuration does not go through and reads are missing
some bytes at the start.

The root cause is yet to be identified. In the meantime, revert this
commit to make sure the flash keeps functioning correctly.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>