clocksource/drivers/timer-ti-dm: Extend pending interrupt ACK for gic
The current implementation for preventing timer interrupts from
breaking suspend fully fixes the issue on am335x but the GIC present
on am437x cannot be directly acked using only the irqchip calls as is
done for am335x but requires an additional step.

Calling irqchip->irq_eoi only writes to the GIC_CPU_EOI register but
for an interrupt to be properly cleared by the GIC, a read from
GIC_CPU_INTACK must come first.  The only place the irq-gic driver
reads this is in the actual interrupt handler so we cannot access it
from the driver.

To get around this, let's map the GIC_CPU_BASE and read the
GIC_CPU_INTACK register ourselves before calling irq_eoi to properly
ack late timer interrupts that show up during suspend on am437x.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
[khilman: migrated from mach-omap2 to clocksource driver]
Signed-off-by: Kevin Hilman <khilman@baylibre.com>