rpmsg-kdrv: ethswitch: add support for mac-only ports
The EthFW has been updated to support mac-only ports - all traffic is sent
only to the host and host uses directed transfer to the port. The all
existing infrastructure and ethswitch protocol are reused for mac-only
mode.

The EthFW indicates presents of virt-maconly port by creating additional
kdrv device named as below:
  "mpu_1_0_ethmac-device-X" where X is device id

And ethswitch protocol extended to indicate mac-only mode to the
j721e-cpsw-virt-mac driver:

 ATTACH(_EXT)
  - features field BIT(2) indicates mac-only mode
  - mac_only_port - new field to indicate CPSWxG ext. port number
configured in mac-only mode. It has to be used for direct DMA data
transfers.

 SET_PROMISC_MODE - new command to control promisc mode (mac-only-caf).

 IPV4_MAC_REGISTER/IPV4_MAC_UNREGISTER - shell not be used in mac-only mode

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>