media: ti: j721e-csi2rx: Drain DMA when dropping frames
Once frames start dropping due to lack of buffers, some data gets stuck
in the DMA pipeline somewhere. So the first DMA transfer after frame
drops gives a partial frame. This is obviously not useful to the
application and will only serve to confuse it. Issue a DMA transaction
to drain that up so that the next frame comes out clean.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>