ti_config_fragments: rpmsg/v8_rpmsg: Enable rpmsg-pru driver
Enable the PRU messaging driver pru-rpmsg (RPMSG_PRU) in both the
ipc.cfg and v8_ipc.cfg so that the driver is built as a module
by default on all applicable TI platforms.  This driver is used
to showcase the remoteproc/rpmsg communication from userspace with
the various PRU or RTU cores present in PRUSS or ICSSG instances.

The driver presents a /dev/rpmsg_pruX (where X is the endpoint
number on PRU) for each rpmsg device that can probe this driver.
This published number needs to be unique across all PRU instances.

The module is built as rpmsg_pru.ko, and the driver name in sysfs
is rpmsg_pru.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>