net: ethernet: ti: j721e-cpsw-virt-mac: add support for mc filtering
The EthFW has been updated to support mc filtering which allows to offload
MC addresses list to EthFW and so allows MC traffic to be forwarded to
Linux Host.

This patch adds corresponding support to virt-mac driver:

- MC filtering is available only if EthFW declares MC_FILTER feature

- The Net core updates netdev MC list from .ndo_set_rx_mode() callback
which can be run with BH disabled. Hence, MC filtering can't be done from
.ndo_set_rx_mode() directly - the workqueue is used to defer MC list
offload which is first copied to the local mc_list under
netif_addr_lock_bh() and then offloaded to EthFW.

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>