arm64: dts: ti: Add J721E CPB OV5640 overlay
J721E EVM boards are not shipped with the OV5640 sensor module. It is a
separate purchase. The module is just one of the possible sensors or
capture boards you can connect. The OV5640 module is not special, it was
just the module with which the capture was tested with.

Having the module in the CPB dts just makes it cumbersome to use other
sensors. Remove its nodes from the dts file and move them to an overlay
so that it is easier to use other sensors as well.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>