arm64: dts: ti: k3-j721e-main: Add PRU system events for virtio
A PRU system event "vring" has been added to each PRU and RTU node
in each of the ICSSG0 and ICSSG1 remote processor subsystems to enable
the virtio/rpmsg communication between MPU and that PRU/RTU core. No
events have been added to the Tx_PRU cores at present. The additions
are done in the base k3-j721e-main.dtsi, and so are inherited by all
the K3 J721E boards.

The PRU system events is the preferred approach over using TI
mailboxes, as it eliminates an external peripheral access from
the PRU/RTU-side, and keeps the interrupt generation internal to
the ICSSG. The difference from MPU would be minimal in using one
versus the other.

Mailboxes can still be used if desired, but currently there is
no support on firmware-side for K3 SoCs to use mailboxes. Either
approach would require that an appropriate firmware image is
loaded/booted on the PRU.

Signed-off-by: Suman Anna <s-anna@ti.com>