net: ethernet: ti: am65-cpsw-qos: iet: allow to set addfragsize
This is required to handle i2208 "ALE IET Express Packet Drops".

>>>
i2208 description:
The issue with ALE is due to CPSW frequency and IET operation with short
express traffic and pre-empted packets that get pre-empted between 60-69
bytes on non-10G capable ports.

If an IET pre-emptible packet get interrupted at 60-69 bytes, the lookup
will occur when the next chunk arrives. The CPSW only gives the ALE 64
bytes from the pre-emptible MAC.

As a result, a short express traffic lookup will start at the end of a 64
byte express traffic, but when the pre-empted queue continues, the
pre-empted traffic will complete the 64 bytes and attempt a lookup for the
pre-empt packet. But this lookup is less that 64 clocks from the express
lookup start, so the express lookup will be aborted(express traffic
dropped) and start the new lookup for the pre-empted traffic.

Workaround(s): During IET negotiation, tell the remote to fragment at 128
bytes.
<<<

Actually, the w/a can't be applied on Host running Linux as it has to be
applied on remote Host, which can be non TI K3 HW:

 DUT1 (ti k3) <------ DUT2 (may be non TI)
 i2208                w/a to apply addfragsize >= 128

This patch adds possibility to configure IET addfragsize through debugfs
parameter "iet_addfragsize" and also sets addfragsize default value to 128.

Note. this interface is subject of change in the future.

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>