HACK: ARM: dts: ti: dra7x: Add GPIO hog for backlight
The origin of this patch is a backport try of:
'backlight: led_bl: Add support for a "enable" GPIO' from jjhiblot@ti.com.
This patch was rejected by upstream and he started to work on a new serie:
"leds: Add control of the voltage/current regulator to the LED core"
until a v7.
But on each iteration, there was still lot of question on needs and rejections.
A complex change for little gains.
So let's just add a GPIO hog for this.

Signed-off-by: Corentin Labbe <clabbe@baylibre.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Reviewed-by: Kevin Hilman <khilman@baylibre.com>