net: ethernet: ti: Fix buffer SRAM overlaps in EMAC mode
Buffer allocation was supposed to follow below layout as per comment in
the prueth_emac_buffer_setup()

[1] |BPOOL0 (64K) |BPOOL1 (64K)|RX_CTX0 (32K)|RX_CTX1 (32K)|

But at the moment tracing prueth_emac_buffer_setup() shows below layout:

[2] |BPOOL0 (64K) |BPOOL1 (64K)|RX_CTX0 (32K)| EMPTY (64K) |RX_CTX1 (32K)|

Additional empty 64K is not required and is a bug. This causes ICSSG0
SRAM reservations to overlap with ICSSG1 and so on. Fix this by dropping
unintended empty 64K hole and this reverting to layout [1].

Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>