remoteproc: debugfs: Optimize the trace va lookup
The commit a987e6b91a5a ("remoteproc: fix trace buffer va initialization")
computes the trace va given the trace device address every time a debugfs
read is done. This can be optimized to perform the lookup only the first
time, and store the value in the va field of the embedded rproc_mem_entry
structure for the debug trace. This restores how the trace va was stored
prior to the above commit in the original trace resource entry handling.

Signed-off-by: Suman Anna <s-anna@ti.com>