mmc: sdhci_am654: Add new compatible for AM62 SoC
The PHY used in the 8 bit instance has been changed to the PHY used in 4
bit instance on AM62 SoC. This implies the PHY configuration required for
both the instances of mmc are similar. Therefore, add a new compatible
for AM62 SoC using the driver data of J721e 4 bit instance.

Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>