arm64: dts: ti: k3-am625-sk: Add ports for dss, HDMI tx and connector
Add and connect output port for DSS (vp2), input and output ports for
the sil9022 HDMI tx and the input port for the HDMI connector.

The dpi output signals from the vp2 of DSS are fed into the sil9022 HDMI
transmitter. These signals are converted into HDMI signals which are
further passed on to the HDMI connector on the board, on which display
can be connected via appropriate HDMI cables.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Tested-by: Darren Etheridge <detheridge@ti.com>