arm64: dts: ti: Introduce k3-am625-sk-lpmdemo
Introduce a new k3-am625-sk-lpmdemo dtb which is a configuration for the
AM625 SK EVM, based on the minimal k3-am625-skeleton to enable checkout
of the minimal set of required devices for low power mode.

This will be extended to enable devices as low power support for them is
validated or added.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>