iio: adc: ti-adc128s052: Fix number of channels when device tree is used
When device_match_data is called - with device tree, of_match list is
looked up to find the data, which by default is 0. So, no matter which
kind of device compatible we use, we match with config 0 which implies
we enable 8 channels even on devices that do not have 8 channels.

Solve it by providing the match data similar to what we do with the ACPI
lookup information.

Fixes: 9e611c9e5a20 ("iio: adc128s052: Add OF match table")
Cc: <stable@vger.kernel.org> # 5.0+
Link: https://lore.kernel.org/linux-iio/20220630230107.13438-1-nm@ti.com/
Signed-off-by: Nishanth Menon <nm@ti.com>