arm64: dts: ti: Add J721S2 CPB OV5640 overlay
J721S2 EVM boards are not shipped with the OV5640 sensor module.
The module is just one of the possible sensors or capture boards
that can be interfaced. The OV5640 module is the module with which
the single camera capture on J721S2 tested with, this commit adds
the overlay for the same and adds corresponding entry in Makefile.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>