arm64: dts: ti: am62x-lp-sk: fix WLAN regulator GPIO and voltage
The schematic uses octal numbering, while linux uses decimal numbering.
Fix the GPIO number in the dts so that the regulator is actually on.

Also the driver seems to want the regulator to be at 3V3 instead of 1V8.
Fix this in the DTS for now, but this might need to be revisited as
the part is actually a level translator converting to 1V8.

Signed-off-by: Anand Gadiyar <gadiyar@ti.com>