arm64: dts: ti: k3-am625-sk-lpmdemo: inherit from k3-am625-sk.dts
The AM625 SK LPM demo DTS file was built up from the skeleton DTS file
to keep a minimal set of tested peripherals with which the SoC could
reliably enter deep sleep mode.

As we now have more peripherals in this list, and we would eventually
like all peripheral drivers to be adapted to allow deep sleep, it makes
more sense to instead start from the regular SK DTS file and disable the
not-yet-adapted modules. As these drivers get updated, we can simply
drop the disabled nodes from this file and we should eventually converge
to the regular DTS file and eliminate the LPM demo dts altogether.

This also ensures that we don't have divergent DTS files as the regular
DTS file gets updated over time.

Signed-off-by: Anand Gadiyar <gadiyar@ti.com>
Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Cc: Vignesh Raghavendra <vigneshr@ti.com>
Tested-by: Georgi Vlaev <g-vlaev@ti.com>