ARM: dts: am335x-bone: Fix ethernetX aliases for beaglebone
When setting ethaddr to some custom MAC address from u-boot on a
beaglebone, the MAC address is not assigned into the correct network
interface. This is due to the aliases being set in am33xx.dtsi,
pointing to cpsw_emac0 and cpsw_emac1.

This patch overrides the ethernet0 and ethernet1 to point at
cpsw_port1 and cpsw_port2.

Tested on a beaglebone rev A4 and with ethaddr unset, the ethernet
interface gets its MAC address from the SoC's e-fuses. With a custom
ethaddr set from u-boot, the ethernet interface has the custom MAC, as
expected.

Signed-off-by: António Oliveira <antonio.oliveira@konsulko.com>
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>