arm64: dts: ti: k3-am62-main: Add ECAP capture nodes
The three Enhanced Capture (ECAP) modules on AM62x can be configured
in either of the following modes: PWM generator or input capture.
This commit adds the input capture nodes to the dts.

To differentiate between the two operating modes, the ECAP nodes have
been renamed as ecapN_pwm and ecapN_capture. The ECAP capture nodes
are disabled by default and can be enabled via an overlay.

The ECAP PWM nodes are currently enabled in main.dtsi, but eventually
get disabled in sk-common and skeleton as the PWM pins are going to
expansion header and might conflict with some functionality. The future
version of main dtsi is expected to disable all such nodes rather than
having this chan of enable and disable. The PWM nodes will also need to
be enabled via an overlay.

Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>