media: ti: j721e-csi2rx: Drain DMA during stream stop
When stopping a stream, the DMA terminate does not currently ensure
a graceful cleanup of the pending data in the fabric.Due to this pending
packet affecting the DMA framing, other contexts that are streaming will
also be interrupted as the CSI_RX_IF expects previous packets to be fully
consumed before any new line, end of frame or start of frame can be sent.
Thus perform ti_csi2rx_drain_dma() to drain any pending data, also update
the log level for the drain DMA failure from dev_warn to dev_dbg during
stop, even when this warning was seen, the next frame was proper this
could be due to the fact that DMACNTX is disabled asynchronously and
the pending data is shorter.

ti_csi2rx_drain_dma() was not properly being invoked now because the
dma->state need not be TI_CSI2RX_DMA_IDLE in all cases of stop stream,
in most cases when the stream was stopped when there were pending buffers
queued the dma->state will be DMA_ACTIVE, but this case also needs cleanup.
This fixes the issue seen during multi-camera streaming when one of the
contexts are stopped, the other contexts are also interrupted.

Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Reviewed-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>